Quase completude da logica de descontos

+++ ex4-pizzaria-base/src/main/java/com/Grupo01/PizzaHUB/Adaptadores/Dados/PedidoRepositoryJDBC.java
@@ -1,12 +1,9 @@
 package com.Grupo01.PizzaHUB.Adaptadores.Dados;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +14,6 @@ import com.Grupo01.PizzaHUB.Dominio.Dados.PedidoRepository;
 import com.Grupo01.PizzaHUB.Dominio.Entidades.Cliente;
 import com.Grupo01.PizzaHUB.Dominio.Entidades.ItemPedido;
 import com.Grupo01.PizzaHUB.Dominio.Entidades.Pedido;
-import com.Grupo01.PizzaHUB.Dominio.Entidades.Pedido.Status;
 
 @Component
 public class PedidoRepositoryJDBC implements PedidoRepository{
@@ -98,12 +94,12 @@ public class PedidoRepositoryJDBC implements PedidoRepository{
         return count != null ? count : 0;
     }
     // Retorna o valor gasto nos últimos 30 dias
-    public float valorGastoUltimos30Dias(String clienteCpf){
+    public double valorGastoUltimos30Dias(String clienteCpf){
         String sql = "SELECT COALESCE(SUM(valorCobrado), 0) FROM pedidos " +
                      "WHERE cliente_cpf = ? AND dataHoraPagamento >= CURRENT_DATE - INTERVAL '30' DAY";
 
         Number soma = jdbcTemplate.queryForObject(sql, Number.class, clienteCpf);
-        return soma == null ? 0f : soma.floatValue();
+        return soma == null ? 0f : soma.doubleValue();
     }
 
     @Override
